requirements.txt
keystone
keystoneauth1
oslo.config
oslo.i18n
oslo.log
neutron-lib
neutron
python-neutronclient
netaddr
requests
